+++ packages/cli/src/index.ts
@@ -8,6 +8,7 @@ import { makeReceiptCommand } from './commands/receipt.js'
 import { makeInitCommand } from './commands/init.js'
 import { makeDiffCommand } from './commands/diff.js'
 import { makeDryRunCommand } from './commands/dry-run.js'
+import { makePolicySnapshotCommand } from './commands/policy-snapshot.js'
 
 const program = new Command()
   .name('txfence')
@@ -21,5 +22,6 @@ program.addCommand(makeReceiptCommand())
 program.addCommand(makeInitCommand())
 program.addCommand(makeDiffCommand())
 program.addCommand(makeDryRunCommand())
+program.addCommand(makePolicySnapshotCommand())
 
 program.parse(process.argv)